+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 Unreleased is the current development version.
 
+- Radiation CLI and diagnostic refinement (#537)
 - Ocean3D CLI and diagnostic refinement (#578)
 - AtmGlobalMean CLI and diagnostic refinement (#587)
 - Tropical cyclones CLI refinements and TC module (#568)